FIX: fixed the editor being unstable after loading the demo data (closes #232)
The problem is that the path is given as a Path object, which is passed as is to QSettings (for saving the recent files list) and in the status.
The first case prevents the editor from even starting *when using several Python versions on the same machine* (I suspect an incompatible pickle),
the second case displayed annoying errors.