FEAT: implemented sas7bdat adapter using Pandas
It is *much* faster than pyreadstat to read the first rows. However, it
does not support offset, so reading the end of large files will be awfully slow.
Even though this is obviously not a perfect solution, I think this is an
improvement because the first rows are often the most interesting ones.